+++ demos/chat/src/components/AudioControls.tsx
@@ -1,0 +1,20 @@
+import BackgroundAudio from "./AudioControls/BackgroundAudio";
+import CharacterAudio from "./AudioControls/CharacterAudio";
+
+const AudioControls = () => {
+  return (
+    <div
+      style={{
+        display: "flex",
+        justifyContent: "space-between",
+        maxWidth: "600px",
+        margin: "0 auto",
+      }}
+    >
+      <BackgroundAudio />
+      <CharacterAudio />
+    </div>
+  );
+};
+
+export default AudioControls;