+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### v7.0.0
+
+- Update and add audio volume functions
+
+### v6.0.2
+
+- Improve character speech volume reactivity by setting it prior to each play speech request
+
 ### v6.0.1
 
 - Exclude tests from build